refactor: type파일 분리
